+++ src/frontends/gtk2/src/immodule.h
@@ -1,7 +1,7 @@
 #pragma once
+#include <X11/Xlib.h>
 #include <gdk/gdk.h>
 #include <gtk/gtkimmodule.h>
-#include <X11/Xlib.h>
 #include <kime_engine.h>
 
 void register_module(GTypeModule *module);